Modify the overview settings to reflect realtime changes
Summary:
Splunk uses time-series data and our current window is 24 hours whilst our polling is 5 minutes.
What this means for us is that all offline/online incidents are grouped in this 24 hour period.
To explain further if a canary was offline and then online it would appear twice in the dataset.
So this Canary would show as online and offline on the Canary Overview page.

We edited the queries to show the latest data on the Canary Splunk App.
This ensures that when the default 24 Window period from the daily_snaphot works as intended.
The new_incidents_devices_token will pull new information into the time series.

We achieved this by removing the older duplicated incidents

Bumped version to 1.2.3

Test Plan:
Tested with Splunk builder (./build_packages.sh -a -A):

Tested built image with a fresh console:
Tests included:
- Adding/removing Canarytokens
- Changing the state of A Canary to live and the offine
- Acking and unacking incidents.

Tests for the incident investigator:
Check that the latest state of the incident shows up.
Tested by:
Acking an incident seeing the result in the incident investigator and then
unacking the incident and making sure it corresponds to what is seen in splunk

Reviewers: #devs, marco

Reviewed By: #devs, marco

Maniphest Tasks: T1769

Differential Revision: http://phabricator.thinkst.com:9000/D2588